Fix calculation of solution path cost

+++ src/main/java/code/generic/QueueingFunctions/UCSQueueingFunction.java
@@ -4,8 +4,6 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.PriorityQueue;
 import java.util.Queue;
-import java.util.Set;
-
 import code.Node;
 
 public class UCSQueueingFunction implements QueueingFunction {
@@ -13,7 +11,7 @@ public class UCSQueueingFunction implements QueueingFunction {
     @Override
     public Queue<Node> apply(Queue<Node> queue, List<Node> nodes) {
         queue.addAll(nodes);
-        PriorityQueue<Node> priorityQueue = new PriorityQueue<>(Comparator.comparingDouble(Node::getPathCost));
+        PriorityQueue<Node> priorityQueue = new PriorityQueue<>(Comparator.comparingInt(Node::getPathCost));
         priorityQueue.addAll(queue);
         return priorityQueue;
     }
